+++ NEWS.md
@@ -15,8 +15,13 @@ subtitle: plm - Linear Models for Panel Data - A set of estimators and tests for
   `cquad` were notified and were provided a patch in Jan 2022). Development
   repository of cquad is <https://github.com/fravale/cquad_dev/>.
 
+### Features:
+* predict.plm: prediction implemented for fixed effects models incl. support for
+  argument 'newdata' and out-of-sample prediction. Help page (`?predict.plm`) 
+  added to specifically explain the prediction for fixed effects models and the 
+  out-of-sample case.
 
-
+***
 
 # plm 2.6-1